+++ tests/src/Advertiser/Repository/DummyStatsRepository.php
@@ -23,6 +23,7 @@ declare(strict_types=1);
 
 namespace Adshares\Tests\Advertiser\Repository;
 
+use Adshares\Adserver\Http\Requests\Filter\FilterCollection;
 use Adshares\Adserver\Models\Campaign;
 use Adshares\Adserver\Models\User;
 use Adshares\Advertiser\Dto\Result\ChartResult;
@@ -44,7 +45,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $data = [
             ['2019-01-01T15:00:00+00:00', 1],
@@ -61,7 +63,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         return $this->fetchView($advertiserId, $resolution, $dateStart, $dateEnd, $campaignId);
     }
@@ -71,7 +74,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $data = [
             ['2019-01-01T15:00:00+00:00', 0.01],
@@ -88,7 +92,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         return $this->fetchView($advertiserId, $resolution, $dateStart, $dateEnd, $campaignId);
     }
@@ -98,7 +103,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $data = [
             ['2019-01-01T15:00:00+00:00', 11],
@@ -115,7 +121,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         return $this->fetchClick($advertiserId, $resolution, $dateStart, $dateEnd, $campaignId);
     }
@@ -125,7 +132,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $data = [
             ['2019-01-01T15:00:00+00:00', 0.02],
@@ -142,7 +150,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $data = [
             ['2019-01-01T15:00:00+00:00', 12],
@@ -159,7 +168,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $data = [
             ['2019-01-01T15:00:00+00:00', 13],
@@ -176,7 +186,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $data = [
             ['2019-01-01T15:00:00+00:00', 14],
@@ -193,7 +204,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $data = [
             ['2019-01-01T15:00:00+00:00', 141],
@@ -210,7 +222,8 @@ class DummyStatsRepository implements StatsRepository
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $data = [
             ['2019-01-01T15:00:00+00:00', 0.03],
@@ -226,7 +239,8 @@ class DummyStatsRepository implements StatsRepository
         ?string $advertiserId,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): DataCollection {
         $user = User::fetchByEmail(self::USER_EMAIL);
 
@@ -290,7 +304,8 @@ class DummyStatsRepository implements StatsRepository
         ?string $advertiserId,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): Total {
         $calculation = new Calculation(1, 1, 1, 1, 1, 1);